// src/shared/types/RentalOfferType.ts
import { RentalCity } from "./RentalCityType.js";
import { Housing } from "./HousingType.js";
import { Amenity } from "./AmenityType.js";
import { Coordinates } from "./CoordinatesType.js";
import { User } from "./UserType.js";


export type RentalOffer = {
  // Наименование. Обязательное. Мин. длин 10 символов, макс. длина 100;
  name: string;
  // Описание предложения. Обязательное. Мин. длина 20 символов, макс. длина 1024 символа;
  description: string;
  // Дата публикации предложения. Обязательное.
  publishedData: Date;
  // Город. Обязательное. Один из шести городов.
  city: RentalCity;
  // Превью изображения. Обязательное. Ссылка на изображение, которое используется в качестве превью;
  previewUrl: string;
  // Фотографии жилья. Обязательное. Список ссылок на фотографии жилья. Всегда 6 фотографий;
  offerImages: string[];
  // Флаг «Премиум». Обязательное. Признак премиальности предложения;
  isPremium: boolean;
  // Флаг «Избранное». Обязательное. Признак того, что предложение принадлежит списку избранных предложений пользователя;
  isFavorite: boolean;
  // Рейтинг.Обязательное.Число от 1 до 5. Допускаются числа с запятой(1 знак после запятой);
  rating: number;
  // Тип жилья.Обязательное.Один из вариантов
  housingType: Housing;
  // Количество комнат.Обязательное.Мин. 1, Макс. 8;
  roomCount: number;
  // Количество гостей. Обязательное. Мин. 1, Макс. 10;
  guestCount: number;
  // Стоимость аренды. Обязательное. Мин. 100, Макс. 100 000;
  price: number;
  // Удобства. Обязательное. Список удобств. Один или несколько вариантов из списка
  amenities: Amenity[];
  // Автор предложения. Обязательное. Ссылка на сущность «Пользователь»;
  author: User;
  // Количество комментариев. Рассчитывается автоматически;
  commentCount: number;
  // Координаты предложения для аренды. Обязательное. Координаты представлены широтой и долготой.
  coordinates: Coordinates;
}